refactor: split handlers + fold lance-path resolution into JobExecutor
Pulls pipeline handlers out of the 1332-line `handlers.rs` into a new
`pipeline_handlers.rs` module; `handlers.rs` now holds only the server-
wide `GET /` dashboard and `GET /health` liveness probe. Shared response
helpers and the generic `ErrorResponse` type move with the pipeline
handlers since nothing else uses them today.

In the jobs layer, `JobExecutor` now owns a `source_paths` map directly
instead of having the server wrap it in a `JobExecutorBundle` and feed
paths through a separate `submit_with_lance_paths` extension trait.
This removes the need to grow a new `submit_with_<lake>_paths` method
every time we add a lake backend (Iceberg, Delta, Paimon, ...).
Dropped: `JobExecutorExt`, `JobExecutorBundle`, `DataSourceInfo`,
`lance_paths_from_sources`, and the helper submit path. `AppState.jobs`
is now `Option<Arc<JobExecutor>>` and handlers call `executor.submit(..)`
directly.
